Rename analyze library
